Widen token counters from u32 to u64
Token accumulator fields across Usage, AgentProgress, TuiState,
SubagentResult, CompactionResult, SessionEntry, and CompactionConfig
are now u64 to avoid theoretical overflow in long sessions. Limit
and ordinal fields (context_window, max_tokens, index) stay u32.

+++ crates/tau-agent/src/agent.rs
@@ -689,10 +689,7 @@ impl Agent {
             return;
         }
         let used = turn_usage.input + turn_usage.cache_read;
-        let limit = self
-            .config
-            .model
-            .context_window
+        let limit = (self.config.model.context_window as u64)
             .saturating_sub(self.config.compaction.reserve_tokens);
         if used > limit {
             let _ = self.run_compaction(CompactionReason::Threshold).await;
